+++ src/main/java/NMQC/C_O_R.java
@@ -20,7 +20,7 @@ import ij.util.*;
 import ij.process.*;
 import ij.measure.*;
 import ij.plugin.filter.PlugInFilter;
-import NMQC.utils.*;
+import utils.*;
 import ij.io.*;
 
 /**
