+++ src/main/java/Tools/Cardiac_Reslicer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package NMQC.Tools;
+package Tools;
 
 import ij.*;
 import ij.gui.*;
